requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
click==8.1.7
schedule==1.2.0
python-dotenv==1.0.0